test(e2e): migrate Interactive UI Snap smoke to Appium
